Add files via upload
Zombie City RPG if my first attempt to create something from scratch on my own. It is based off of a board game that I created called Zombie City.  This RPG holds most of the common elements of an RPG, Health, Mana, Weapons and spells. There is a final boss as well. All of the art and sound are original.